* Fix a problem with compilation of libffi that I introduced in my previous checkin * Explictly add '-fexceptions' to the compiler arguments (probably not needed) * Fix some 64-bit issues (NSInteger vs. int issues) * More serious: code sometimes looked at PyObjCObject fields without checking that   the object really is a PyObjCObject, even when objects can PyObjCClass instances.   This happens to work fine on all platforms but PPC64, but that's accidental. * Enable '@try{' instead of 'NSDURING'
All of this is part of trying to get pyobjc-core going on a PPC64 system. Sadly
enough I now get a fatal error from the ObjC runtime because of an uncaught
exception, even though code is passing through an @try block that should catch
the exception.

+++ pyobjc/pyobjc-core/Modules/objc/test/testbndl.m
@@ -1663,7 +1663,7 @@ static 	char buf[1024];
 + (NSData*)fetchData:(NSCoder*)coder
 {
 	void* data;
-	unsigned length;
+	NSUInteger length;
 
 	data = [coder decodeBytesWithReturnedLength:&length];
 	return [NSData dataWithBytes:data length:length];
